begin on refactoring
