+++ src/hash/formats/providers/libxml.rs
@@ -1,0 +1,7 @@
+pub struct LibXML;
+
+impl super::XMLProvider for LibXML {
+    fn find_xpath(source: &str, xpath: &str) -> Option<String> {
+        unimplemented!()
+    }
+}
