+++ src/hash/formats/providers/mod.rs
@@ -1,5 +1,16 @@
 //! Providers for xml parsing
 
-#[cfg(feature = "libxml")]
-mod libxml;
+pub trait XMLProvider {
+    fn find_xpath(source: &str, xpath: &str) -> Option<String>;
+}
+
+cfg_if::cfg_if! {
+    if #[cfg(feature = "libxml")] {
+        mod libxml;
+        pub use libxml::LibXML as Provider;
+    } else {
+        pub use other::Other as Provider;
+    }
+}
+
 mod other;
